# backend_and_api/openapi.yaml
openapi: 3.0.2
info:
  description: This is the API for traffic wise based on OpenAPI 3.0 specification..
  title: Traffic Wise
  version: "1.0"
servers:
- url: /v3
tags:
- description: Find routes between your starting place and destination.
  name: Routes
- description: Convert between latlong and addresses.
  name: Points
- description: Custom preferences
  name: Preferences
paths:
  /prefer:
    description: Set custom preferences for client.
    put:
      operationId: setPreferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Preferences'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Preferences'
        description: JSON Settings
        required: true
      responses:
        "200":
          description: Confirm Preferences are set
      tags:
      - Preferences
  /addresses:
    description: Get address from backend.Latitude and Longitude
    get:
      description: get address in different format
      operationId: getAddressFromLatLong
      parameters:
      - explode: true
        in: query
        name: latitude
        required: true
        schema:
          example: "45.424721"
          type: string
        style: form
      - explode: true
        in: query
        name: longitude
        required: true
        schema:
          example: "-75.695000"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
            application/x-www-form-urlencoded:
              schema:
                type: string
          description: All good
      tags:
      - Points
  /latlong:
    get:
      description: get Latitude and Longitude from backend.address
      operationId: getLatLongFromAddress
      parameters:
      - explode: true
        in: query
        name: address
        required: true
        schema:
          example: 8 Sheldrake Drive Kanata
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coords'
            application/xml:
              schema:
                $ref: '#/components/schemas/Coords'
          description: All good
      tags:
      - Points
  /saferoute:
    post:
      description: Get the routes
      operationId: getSafeRoute
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SafeRoutes'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SafeRoutes'
        description: JSON Settings
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SingleResp'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/SingleResp'
                type: array
          description: Successful return of routes
      tags:
      - Routes
components:
  schemas:
    SafeRoutesResp:
      items:
        $ref: '#/components/schemas/SingleResp'
      type: array
    SingleResp:
      example:
        routes:
        - routes
        - routes
        traveltime: 6.0274563
        distance: 1.4658129
        risk: 0.8008282
      properties:
        routes:
          items:
            type: string
          title: routes
          type: array
        risk:
          format: float
          title: risk
          type: number
        traveltime:
          format: float
          title: traveltime
          type: number
        distance:
          format: float
          title: distance
          type: number
      title: SingleResp
      type: object
    Coords:
      example:
        latitude: latitude
        longitude: longitude
      properties:
        latitude:
          title: latitude
          type: string
        longitude:
          title: longitude
          type: string
      title: Coords
      type: object
    Preferences:
      example:
        vehicletype: CAR
        environment: NORMAL
        PathChoices:
          distance: 0.14658129
          safety: 0.08008282
          time_of_day: 0.6027456
      properties:
        vehicletype:
          default: CAR
          enum:
          - PEDESTRIAN
          - BICYCLE
          - MOTORCYCLE
          - CAR
          - TRUCK
          - OTHER
          title: vehicletype
          type: string
        environment:
          default: NORMAL
          enum:
          - IDEAL
          - NORMAL
          - ABNORMAL
          - DEVIANT
          - TROUBLESOME
          - EXTREME
          title: environment
          type: string
        PathChoices:
          $ref: '#/components/schemas/Preferences_PathChoices'
      title: Preferences
      type: object
    SafeRoutes:
      example:
        path_count: 0.40041409523050575
        StartTime:
          time_of_day: "1230"
          day_of_week: MONDAY
        StartEnd:
          destination_address: "40 Merivale Road, Ottawa"
          start_address: 8 Sheldrake Drive Kanata
      properties:
        StartEnd:
          $ref: '#/components/schemas/SafeRoutes_StartEnd'
        StartTime:
          $ref: '#/components/schemas/SafeRoutes_StartTime'
        path_count:
          maximum: 5
          minimum: 0
          title: path_count
          type: number
      title: SafeRoutes
      type: object
    Preferences_PathChoices:
      example:
        distance: 0.14658129
        safety: 0.08008282
        time_of_day: 0.6027456
      properties:
        safety:
          default: 0.33
          format: float
          maximum: 1
          minimum: 0
          title: safety
          type: number
        time_of_day:
          default: 0.33
          format: float
          maximum: 1
          minimum: 0
          title: time_of_day
          type: number
        distance:
          default: 0.33
          format: float
          maximum: 1
          minimum: 0
          title: distance
          type: number
      title: Preferences_PathChoices
      type: object
    SafeRoutes_StartEnd:
      example:
        destination_address: "40 Merivale Road, Ottawa"
        start_address: 8 Sheldrake Drive Kanata
      properties:
        start_address:
          example: 8 Sheldrake Drive Kanata
          title: start_address
          type: string
        destination_address:
          example: "40 Merivale Road, Ottawa"
          title: destination_address
          type: string
      title: SafeRoutes_StartEnd
      type: object
    SafeRoutes_StartTime:
      example:
        time_of_day: "1230"
        day_of_week: MONDAY
      properties:
        day_of_week:
          default: MONDAY
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
          title: day_of_week
          type: string
        time_of_day:
          example: "1230"
          title: time_of_day
          type: string
      title: SafeRoutes_StartTime
      type: object