# backend_and_api/src/openapi_server/models/preferences.py
# coding: utf-8

"""
    Traffic Wise

    This is the API for traffic wise based on OpenAPI 3.0 specification..

    The version of the OpenAPI document: 1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List, Optional
from openapi_server.models.preferences_path_choices import PreferencesPathChoices
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class Preferences(BaseModel):
    """
    Preferences
    """ # noqa: E501
    vehicletype: Optional[StrictStr] = 'CAR'
    environment: Optional[StrictStr] = 'NORMAL'
    path_choices: Optional[PreferencesPathChoices] = Field(default=None, alias="PathChoices")
    __properties: ClassVar[List[str]] = ["vehicletype", "environment", "PathChoices"]

    @field_validator('vehicletype')
    def vehicletype_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value
        if value not in ('PEDESTRIAN', 'BICYCLE', 'MOTORCYCLE', 'CAR', 'TRUCK', 'OTHER'):
            raise ValueError("must be one of enum values ('PEDESTRIAN', 'BICYCLE', 'MOTORCYCLE', 'CAR', 'TRUCK', 'OTHER')")
        return value

    @field_validator('environment')
    def environment_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value
        if value not in ('IDEAL', 'NORMAL', 'ABNORMAL', 'DEVIANT', 'TROUBLESOME', 'EXTREME'):
            raise ValueError("must be one of enum values ('IDEAL', 'NORMAL', 'ABNORMAL', 'DEVIANT', 'TROUBLESOME', 'EXTREME')")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        print(self)
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of Preferences from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of path_choices
        if self.path_choices:
            _dict['PathChoices'] = self.path_choices.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of Preferences from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "vehicletype": obj.get("vehicletype") if obj.get("vehicletype") is not None else 'CAR',
            "environment": obj.get("environment") if obj.get("environment") is not None else 'NORMAL',
            "PathChoices": PreferencesPathChoices.from_dict(obj.get("PathChoices")) if obj.get("PathChoices") is not None else None
        })
        return _obj