# backend_and_api/src/openapi_server/models/safe_routes_resp.py
# coding: utf-8

"""
    Traffic Wise

    This is the API for traffic wise based on OpenAPI 3.0 specification..

    The version of the OpenAPI document: 1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


import pprint
import re  # noqa: F401
import json


from pydantic import BaseModel, ConfigDict, StrictFloat, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List, Optional, Union
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class SafeRoutesResp(BaseModel):
    """
    SafeRoutesResp
    """ # noqa: E501
    routes: Optional[List[StrictStr]] = None
    risk: Optional[Union[StrictFloat, StrictInt]] = None
    traveltime: Optional[Union[StrictFloat, StrictInt]] = None
    distance: Optional[Union[StrictFloat, StrictInt]] = None
    __properties: ClassVar[List[str]] = ["routes", "risk", "traveltime", "distance"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of SafeRoutesResp from backend.a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from backend.calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of SafeRoutesResp from backend.a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "routes": obj.get("routes"),
            "risk": obj.get("risk"),
            "traveltime": obj.get("traveltime"),
            "distance": obj.get("distance")
        })
        return _obj